Recurring-task-history

+++ backend/database/recurringTask.go
@@ -3,6 +3,7 @@ package database
 import (
 	"context"
 	"database/sql"
+	"errors"
 	"time"
 
 	"github.com/cockroachdb/cockroach-go/v2/crdb"
@@ -10,15 +11,16 @@ import (
 )
 
 type RecurringTask struct {
-	Id          uuid.UUID `json:"id"`
-	Name        string    `json:"name"`
-	Description string    `json:"desc"`
-	Start       time.Time `json:"start"`
-	Ending      time.Time `json:"ending"`
-	Interval    int       `json:"interval"`
-	CreatedAt   time.Time `json:"createdAt"`
-	UpdatedAt   time.Time `json:"updatedAt"`
-	UserId      uuid.UUID `json:"userId"`
+	Id          uuid.UUID              `json:"id"`
+	Name        string                 `json:"name"`
+	Description string                 `json:"desc"`
+	Start       time.Time              `json:"start"`
+	Ending      *time.Time             `json:"ending"`
+	Interval    int                    `json:"interval"`
+	CreatedAt   time.Time              `json:"createdAt"`
+	UpdatedAt   time.Time              `json:"updatedAt"`
+	UserId      uuid.UUID              `json:"userId"`
+	History     []RecurringTaskHistory `json:"history"`
 }
 
 func (t *RecurringTask) merge(s *RecurringTask) {
@@ -31,12 +33,31 @@ func (t *RecurringTask) merge(s *RecurringTask) {
 	if s.Start != (time.Time{}) {
 		t.Start = s.Start
 	}
-	if s.Ending != (time.Time{}) {
+	if s.Ending != nil {
 		t.Start = s.Start
 	}
 	if s.Interval != 0 {
 		t.Interval = s.Interval
 	}
+}
+
+func (s *DBService) CheckRecurringTaskExists(id uuid.UUID, uid uuid.UUID) bool {
+	res, err := s.db.Query(
+		"SELECT id FROM recurring_tasks WHERE id=$1 AND user_id=$2",
+		id,
+		uid,
+	)
+	if err != nil {
+		return false
+	}
+	defer res.Close()
+	res.Next()
+	var Id uuid.UUID
+
+	if err := res.Scan(&Id); err != nil {
+		return false
+	}
+	return true
 }
 
 func (s *DBService) GetRecurringTaskById(id uuid.UUID, uid uuid.UUID) (*RecurringTask, error) {
@@ -56,7 +77,7 @@ func (s *DBService) GetRecurringTaskById(id uuid.UUID, uid uuid.UUID) (*Recurrin
 	var name string
 	var desc string
 	var start time.Time
-	var ending time.Time
+	var ending sql.NullTime
 	var interval int
 	var createdAt time.Time
 	var updatedAt time.Time
@@ -65,8 +86,13 @@ func (s *DBService) GetRecurringTaskById(id uuid.UUID, uid uuid.UUID) (*Recurrin
 	if err := res.Scan(&tId, &name, &desc, &start, &ending, &interval, &createdAt, &updatedAt, &userId); err != nil {
 		return nil, err
 	}
-	task := RecurringTask{Id: tId, Name: name, Description: desc, Start: start, Ending: ending, Interval: interval, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId}
-	return &task, nil
+	var end *time.Time = nil
+	if ending.Valid {
+		end = &ending.Time
+	}
+	task := RecurringTask{Id: tId, Name: name, Description: desc, Start: start, Ending: end, Interval: interval, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId}
+	task.History, err = s.GetRecurringTasksHistory(task.Id, task.UserId)
+	return &task, err
 }
 
 func (s *DBService) GetRecurringTasksByUser(id uuid.UUID) ([]RecurringTask, error) {
@@ -84,7 +110,7 @@ func (s *DBService) GetRecurringTasksByUser(id uuid.UUID) ([]RecurringTask, erro
 		var name string
 		var desc string
 		var start time.Time
-		var ending time.Time
+		var ending sql.NullTime
 		var interval int
 		var createdAt time.Time
 		var updatedAt time.Time
@@ -93,7 +119,15 @@ func (s *DBService) GetRecurringTasksByUser(id uuid.UUID) ([]RecurringTask, erro
 		if err := res.Scan(&tId, &name, &desc, &start, &ending, &interval, &createdAt, &updatedAt, &userId); err != nil {
 			return nil, err
 		}
-		task := RecurringTask{Id: tId, Name: name, Description: desc, Start: start, Ending: ending, Interval: interval, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId}
+		var end *time.Time = nil
+		if ending.Valid {
+			end = &ending.Time
+		}
+		task := RecurringTask{Id: tId, Name: name, Description: desc, Start: start, Ending: end, Interval: interval, CreatedAt: createdAt, UpdatedAt: updatedAt, UserId: userId}
+		task.History, err = s.GetRecurringTasksHistory(task.Id, task.UserId)
+		if err != nil {
+			return nil, err
+		}
 		tasks = append(tasks, task)
 	}
 	return tasks, nil
@@ -140,6 +174,14 @@ func (s *DBService) UpdateRecurringTask(reqTask RecurringTask) error {
 }
 
 func (s *DBService) DeleteRecurringTask(id uuid.UUID, userId uuid.UUID) error {
+	task, err := s.GetRecurringTaskById(id, userId)
+	if err != nil {
+		return err
+	}
+	if task.Id != id {
+		return errors.New("unauthorized deletion")
+	}
+	s.DeleteCompleteRecurringTaskHistory(id, userId)
 	res, err := s.db.Query(
 		"DELETE FROM recurring_tasks WHERE id = $1 AND user_id=$2",
 		id,
